Use the correct labels when training the classifier
The current code used the same labels of the first batch for all
data points if more then 100 data points were present in the
data added to the map.

Due to the distribution of labels this appears to not change the
outcome of AUC / ROC values, however, it may change the shape of
the resulting map to some extent. The stochastic nature of the
classifier means that such wrongly labelled data points may not
ruin the classifier output.